Fixed some bugs in the ALU that were preventing simulation from working.
Blew away most of the ALU because it had a bunch of old and krufty stuff that
was causing errors during compilation.  Also had to hack a bit to get the
includes directives to work (I hate that these are so simulator dependent).  It
sims with Icarus right now.  Only a third of the test cases are passing right
now, but that's probably because none of the overflow logic has been implemented
correctly yet.