Modifying the processor simulator to allow for multiple modes.
After a bit of thinking, it dawned on me that a simple way to allow for multiple
different simulations to be run would be to just select them using a case
statement.  So, right now there are only three choices, but as time goes on and
new simulations are added into the mix, this should allow them to grow fairly
easily.  There might actually end up being enough overlap that the case
statement can be thrown out and a generic method applied (i.e., all commands are
the same except for something like a $mode variable that).  For now, I'm going
to do it this way until the whole thing is up and running.